refactor: move old slider to previewer
