+++ src/sections/edit-collection-featured-items/featured-items-form/preview-carousel/slider/Slider.module.scss
